+++ README.md
@@ -30,6 +30,7 @@ Init DB added user `MrRobot` with `fsociety` as password.
 * [ ] Add tests
 * [X] Create Access Level [Admin, Visitor]
 * [X] Create endpoint to add, update, and delete users
-* [ ] Allow None Admin to change their fullname, email and password 
+* [ ] Allow None Admin to change their fullname, email and password
+* [ ] Add input strong password and email validation with Pydataic
